requirements.txt
streamlit
supabase
gotrue
python-dotenv
pydantic
requests
fastapi
uvicorn
pandas
